requirements.txt
streamlit
torch
torchvision
numpy
pandas
scikit-learn
joblib
sentence-transformers
tqdm
Pillow
